requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
litellm>=1.57.0
pyyaml>=6.0
pydantic>=2.0
rich>=13.0
click>=8.0
aiofiles>=24.0
httpx>=0.28.0
jinja2>=3.1
python-multipart>=0.0.18
